DragonDosTools can now write files to CAS tape filesystems.
Extent the DragonDosTools GET command to write files from virtual diskettes
to a CAS tape filesystem.  If the local filename ends with a .CAS
extension, then the file will be written to a CAS file using that name,
and an optional fourth argument specifies the name of the file within
the CAS filesystem.