Add simple program to display HFE disk information.
Hfediskinfo will print information from the header of a HFE virtual diskette
image and information about each track and sector.  More richness and
flexibility in the output might be added in the future.

Made HfeTrack public and accessible from HfeDisk to allow for the necessary
information about diskette tracks to be extracted.